requirements.txt
asgiref==3.8.1
certifi==2025.4.26
crispy-bootstrap4==2024.10
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
geopandas==1.0.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyogrio==0.10.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
shapely==2.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv~=1.1.0
gunicorn
folium==0.15.1psutil==5.9.8
